Add basic tests
Assume the test are run by pytest.

The tests are minnimal, and are meant to check that the interface
to the XSPEC model code is working, not that the model code
itself is working as expected.

There is a convolution-test in the ui version that is currently
excluded as it crashes Python.