Move FORTRAN build into setuptools
Rather than manually compiling the FORTRAN code at run time, this moves
the logic into the setuptools/distutils build_ext command via sub-classing.
It also lets the compiler object do the actual compilation, rather than
doing it manually.

This is better than before, but there's still a lot of hard-coded knowledge
(which in itself isn't terrible here), as well as the fact that it still
doesn't stop the FORTRAN code from always being rebuilt, even when it
doesn't need to be.